Reduce re-rendering of the Chart component on mount/update.
This commit reduces the amount of work involved in chart rendering by:
- Doing an initial empty render rather than a full render. The first
render is almost always incorrect anyway, as the chart needs to know
what space it will be allocated in the page to determine the correct
axis sizing to use.
- Avoiding two updates on every resize. Resize used to update state
and trigger a callback that updated state again, resulting in two
renders. This has been refactored to perform all the state changes
in a single pass.